bussiness features implemented

+++ backend/src/main/java/com/example/repository/OrderRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface OrderRepository
         extends JpaRepository<Order, Integer> {
 
-    List<Order> findByCustomerID(int customerID);
+    List<Order> findByCustomerID(Long customerID);
 
     List<Order> findByStatus(String status);
 }
